refactor(tests/planner): split 5 large test files into 16 focused files along concern boundaries
Split 5 monolithic test files (5,013 lines, 222 tests) into 16 focused
files organized by the callable or concern they test. Move shared test
helpers (_make_refined_wps, _make_manifest, _make_minimal_output_dir)
from module-level into conftest.py as public helpers.

Split mapping:
- test_consolidation.py  → test_consolidation_{merge,writeback,cycles}.py
- test_merge.py          → test_merge_{files,items,snapshot,refine}.py
- test_validation.py     → test_validation_{core,checks,discovery}.py
- test_manifests.py      → test_manifests_{builders,finalize,expansion}.py
- test_planner.py        → test_planner_{api,validation_gate,contracts}.py

Update tests/planner/CLAUDE.md with the new file inventory and
.autoskillit/test-source-map.json with the new test paths for the
aggressive filter mode.
