--- tests/planner/test_planner.py
@@ -1,718 +1,0 @@
-"""Tests for the planner L1 subpackage scaffold."""
-
-from __future__ import annotations
-
-import inspect
-import json
-from pathlib import Path
-
-import pytest
-
-pytestmark = [pytest.mark.layer("planner"), pytest.mark.small, pytest.mark.feature("planner")]
-
-
-def test_planner_package_importable() -> None:
-    import autoskillit.planner  # noqa: F401
-
-
-def test_planner_all_exports() -> None:
-    from autoskillit.planner import __all__
-
-    assert set(__all__) == {
-        "ASSIGN_RESULT_FILE_RE",
-        "DELIVERABLE_BOUNDS",
-        "PHASE_RESULT_FILE_RE",
-        "WP_RESULT_FILE_RE",
-        "build_phase_assignment_manifest",
-        "build_phase_wp_manifest",
-        "compile_plan",
-        "consolidate_wps",
-        "create_run_dir",
-        "expand_assignments",
-        "expand_wps",
-        "finalize_wp_manifest",
-        "validate_plan",
-        "PlannerManifest",
-        "PlannerManifestItem",
-        "merge_files",
-        "merge_tier_results",
-        "extract_item",
-        "replace_item",
-        "build_plan_snapshot",
-        "merge_refined_assignments",
-        "PlanDocument",
-        "PhaseShort",
-        "PhaseElaborated",
-        "AssignmentShort",
-        "AssignmentElaborated",
-        "WPShort",
-        "WPElaborated",
-        "ASSIGNMENT_REQUIRED_KEYS",
-        "PHASE_REQUIRED_KEYS",
-        "WP_REQUIRED_KEYS",
-        "resolve_wp_id",
-        "validate_refined_assignments",
-        "validate_refined_plan",
-        "ValidationFinding",
-        "resolve_task_input",
-        "DiscoveryResult",
-        "discover_tier_files",
-        "load_lifecycle_registry",
-        "record_lifecycle_event",
-    }
-
-
-def test_sequential_state_machine_removed() -> None:
-    """Guard: sequential state machine functions must not exist in manifests."""
-    import autoskillit.planner.manifests as m
-
-    for name in (
-        "check_remaining",
-        "build_assignment_manifest",
-        "build_wp_manifest",
-        "_backstop_wp_index",
-    ):
-        assert not hasattr(m, name), f"{name} still exists — should have been removed"
-
-
-def test_create_run_dir_requires_temp_dir_argument() -> None:
-    """create_run_dir must accept temp_dir as explicit parameter, not read os.environ."""
-    from autoskillit.planner import create_run_dir
-
-    sig = inspect.signature(create_run_dir)
-    assert "temp_dir" in sig.parameters, (
-        "create_run_dir must accept temp_dir as an explicit keyword argument"
-    )
-    assert sig.parameters["temp_dir"].default is inspect.Parameter.empty, (
-        "temp_dir must be a required parameter (no default)"
-    )
-
-
-def test_create_run_dir_does_not_read_environ(tmp_path, monkeypatch) -> None:
-    """create_run_dir must not depend on AUTOSKILLIT_TEMP env var."""
-    from autoskillit.planner import create_run_dir
-
-    monkeypatch.delenv("AUTOSKILLIT_TEMP", raising=False)
-    result = create_run_dir(temp_dir=str(tmp_path))
-    assert result["planner_dir"]
-    assert Path(result["planner_dir"]).exists()
-
-
-def test_create_run_dir_creates_timestamped_directory(tmp_path) -> None:
-    from autoskillit.planner import create_run_dir
-
-    result = create_run_dir(temp_dir=str(tmp_path))
-
-    assert "planner_dir" in result
-    planner_dir = Path(result["planner_dir"])
-    assert planner_dir.exists()
-    assert planner_dir.parent.name == "planner"
-    assert planner_dir.name.startswith("run-")
-    assert (planner_dir / "phases").is_dir()
-    assert (planner_dir / "assignments").is_dir()
-    assert (planner_dir / "work_packages").is_dir()
-
-
-def test_create_run_dir_unique_across_calls(tmp_path) -> None:
-    from autoskillit.planner import create_run_dir
-
-    r1 = create_run_dir(temp_dir=str(tmp_path))
-    r2 = create_run_dir(temp_dir=str(tmp_path))
-
-    assert r1["planner_dir"] != r2["planner_dir"]
-    planner_dir2 = Path(r2["planner_dir"])
-    assert planner_dir2.exists()
-    assert planner_dir2.parent.name == "planner"
-    assert planner_dir2.name.startswith("run-")
-    assert (planner_dir2 / "phases").is_dir()
-    assert (planner_dir2 / "assignments").is_dir()
-    assert (planner_dir2 / "work_packages").is_dir()
-
-
-def test_create_run_dir_empty_string_raises_value_error() -> None:
-    from autoskillit.planner import create_run_dir
-
-    with pytest.raises(ValueError, match="non-empty"):
-        create_run_dir("")
-
-
-def test_planner_feature_skill_categories() -> None:
-    from autoskillit.core.types._type_constants import FEATURE_REGISTRY
-
-    assert FEATURE_REGISTRY["planner"].skill_categories == frozenset({"planner"})
-
-
-# --- T6: New callables importable ---
-
-
-def test_expand_assignments_importable() -> None:
-    from autoskillit.planner.manifests import expand_assignments
-
-    assert callable(expand_assignments)
-
-
-def test_expand_wps_importable() -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    assert callable(expand_wps)
-
-
-# --- T7: New callables functional tests ---
-
-
-def test_merge_tier_results_globs_and_merges(tmp_path) -> None:
-    from autoskillit.planner.merge import merge_tier_results
-
-    results_dir = tmp_path / "phases"
-    results_dir.mkdir()
-    for pid in ("P1", "P2"):
-        (results_dir / f"{pid}_result.json").write_text(
-            json.dumps({"id": pid, "name": f"Phase {pid}", "ordering": int(pid[1])})
-        )
-    out = tmp_path / "combined.json"
-    result = merge_tier_results(str(results_dir), str(out), "phases")
-    assert result["item_count"] == "2"
-    assert out.exists()
-    merged = json.loads(out.read_text())
-    merged_ids = {item["id"] for item in merged["phases"]}
-    assert merged_ids == {"P1", "P2"}
-
-
-def test_expand_assignments_creates_contexts(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_assignments
-
-    refined = tmp_path / "refined_plan.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "phases": [
-                    {
-                        "id": "P1",
-                        "name": "Phase 1",
-                        "ordering": 1,
-                        "assignments_preview": [{"name": "A1"}, {"name": "A2"}],
-                    },
-                ]
-            }
-        )
-    )
-    result = expand_assignments(str(refined), str(tmp_path))
-    assert "manifest_path" in result
-    assert "context_paths" in result
-    assert "P1" in result.get("item_ids", "")
-
-
-def test_expand_wps_creates_contexts(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    refined = tmp_path / "refined_assignments.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "assignments": [
-                    {
-                        "phase_id": "P1",
-                        "phase_name": "Phase 1",
-                        "phase_number": 1,
-                        "assignment_number": 1,
-                        "proposed_work_packages": [
-                            {"id": "P1-A1-WP1", "name": "WP1", "scope": "scope1"},
-                        ],
-                    },
-                ]
-            }
-        )
-    )
-    result = expand_wps(str(refined), str(tmp_path))
-    assert "manifest_path" in result
-    assert "context_paths" in result
-    assert "P1" in result.get("item_ids", "")
-
-
-# --- Validation gate tests (ARCH-010) ---
-
-
-def test_validate_refined_assignments_rejects_empty_id() -> None:
-    from autoskillit.planner.schema import validate_refined_assignments
-
-    data = {
-        "assignments": [
-            {
-                "id": "P1-A1",
-                "phase_id": "P1",
-                "phase_number": 1,
-                "assignment_number": 1,
-                "proposed_work_packages": [{"name": "No ID", "scope": "s"}],
-            }
-        ]
-    }
-    with pytest.raises(ValueError, match="neither 'id' nor 'id_suffix'"):
-        validate_refined_assignments(data)
-
-
-def test_validate_refined_assignments_resolves_id_suffix() -> None:
-    from autoskillit.planner.schema import validate_refined_assignments
-
-    data = {
-        "assignments": [
-            {
-                "id": "P1-A1",
-                "phase_id": "P1",
-                "phase_number": 1,
-                "assignment_number": 1,
-                "proposed_work_packages": [
-                    {"id_suffix": "WP1", "name": "First", "scope": "s1"},
-                ],
-            }
-        ]
-    }
-    result = validate_refined_assignments(data)
-    wp = result["assignments"][0]["proposed_work_packages"][0]
-    assert wp["id"] == "P1-A1-WP1"
-
-
-def test_validate_refined_assignments_rejects_unresolvable_assign_id() -> None:
-    from autoskillit.planner.schema import validate_refined_assignments
-
-    data = {
-        "assignments": [
-            {
-                "phase_name": "Phase 1",
-                "proposed_work_packages": [{"id_suffix": "WP1", "name": "X"}],
-            }
-        ]
-    }
-    with pytest.raises(ValueError, match="assignment.*resolvable id"):
-        validate_refined_assignments(data)
-
-
-def test_validate_refined_plan_rejects_empty_phase_id() -> None:
-    from autoskillit.planner.schema import validate_refined_plan
-
-    data = {
-        "phases": [
-            {
-                "id": "",
-                "name": "Bad Phase",
-                "assignments_preview": [],
-            }
-        ]
-    }
-    with pytest.raises(ValueError, match="empty.*id"):
-        validate_refined_plan(data)
-
-
-def test_expand_assignments_rejects_empty_preview_ids(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_assignments
-
-    refined = tmp_path / "refined_plan.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "phases": [
-                    {
-                        "id": "P1",
-                        "name": "Phase 1",
-                        "assignments_preview": [{"name": "A1"}],
-                    }
-                ]
-            }
-        )
-    )
-    result = expand_assignments(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    ids = manifest["items"][0]["metadata"]["assignment_ids"]
-    assert all(id_val != "" for id_val in ids)
-
-
-def test_resolve_wp_id_from_id() -> None:
-    from autoskillit.planner.schema import resolve_wp_id
-
-    assert resolve_wp_id({"id": "P1-A1-WP1"}, "P1-A1") == "P1-A1-WP1"
-
-
-def test_resolve_wp_id_from_suffix() -> None:
-    from autoskillit.planner.schema import resolve_wp_id
-
-    assert resolve_wp_id({"id_suffix": "WP2"}, "P1-A1") == "P1-A1-WP2"
-
-
-def test_resolve_wp_id_raises_on_neither() -> None:
-    from autoskillit.planner.schema import resolve_wp_id
-
-    with pytest.raises(ValueError, match="neither 'id' nor 'id_suffix'"):
-        resolve_wp_id({"name": "No ID"}, "P1-A1")
-
-
-def test_resolve_wp_id_rejects_empty_string_id() -> None:
-    from autoskillit.planner.schema import resolve_wp_id
-
-    assert resolve_wp_id({"id": "", "id_suffix": "WP1"}, "P1-A1") == "P1-A1-WP1"
-
-
-def test_expand_wps_resolves_id_suffix(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    refined = tmp_path / "refined_assignments.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "assignments": [
-                    {
-                        "id": "P1-A1",
-                        "phase_id": "P1",
-                        "phase_number": 1,
-                        "assignment_number": 1,
-                        "proposed_work_packages": [
-                            {"id_suffix": "WP1", "name": "First", "scope": "s1"},
-                            {"id_suffix": "WP2", "name": "Second", "scope": "s2"},
-                        ],
-                    }
-                ]
-            }
-        )
-    )
-    result = expand_wps(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    wp_ids = manifest["items"][0]["metadata"]["wp_ids"]
-    assert wp_ids == ["P1-A1-WP1", "P1-A1-WP2"]
-
-
-def test_expand_wps_synthesizes_assign_id_from_numbers(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    refined = tmp_path / "refined_assignments.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "assignments": [
-                    {
-                        "phase_number": 2,
-                        "assignment_number": 3,
-                        "proposed_work_packages": [
-                            {"id_suffix": "WP1", "name": "WP", "scope": "s"},
-                        ],
-                    }
-                ]
-            }
-        )
-    )
-    result = expand_wps(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    wp_ids = manifest["items"][0]["metadata"]["wp_ids"]
-    assert wp_ids == ["P2-A3-WP1"]
-
-
-def test_expand_wps_raises_on_missing_id_and_suffix(tmp_path) -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    refined = tmp_path / "refined_assignments.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "assignments": [
-                    {
-                        "id": "P1-A1",
-                        "phase_id": "P1",
-                        "phase_number": 1,
-                        "assignment_number": 1,
-                        "proposed_work_packages": [
-                            {"name": "No ID WP", "scope": "s1"},
-                        ],
-                    }
-                ]
-            }
-        )
-    )
-    with pytest.raises(ValueError, match="neither 'id' nor 'id_suffix'"):
-        expand_wps(str(refined), str(tmp_path))
-
-
-def test_merge_files_does_not_use_atomic_write(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    """merge_files must write directly under flock, not via atomic_write/os.replace."""
-    from unittest.mock import MagicMock
-
-    from autoskillit.planner.merge import merge_files
-
-    spy = MagicMock(side_effect=AssertionError("atomic_write must not be called"))
-    monkeypatch.setattr("autoskillit.planner.merge.write_versioned_json", spy)
-
-    results_dir = tmp_path / "phases"
-    results_dir.mkdir()
-    (results_dir / "P1_result.json").write_text(
-        json.dumps({"id": "P1", "name": "Phase 1", "ordering": 1})
-    )
-
-    out = tmp_path / "combined.json"
-    result = merge_files([str(results_dir / "P1_result.json")], str(out), "phases")
-    assert result["item_count"] == "1"
-    merged = json.loads(out.read_text())
-    assert merged["schema_version"] == 1
-    assert len(merged["phases"]) == 1
-    spy.assert_not_called()
-
-
-def test_replace_item_does_not_use_atomic_write(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    """replace_item must write directly under flock, not via atomic_write/os.replace."""
-    from unittest.mock import MagicMock
-
-    from autoskillit.planner.merge import replace_item
-
-    spy = MagicMock(side_effect=AssertionError("atomic_write must not be called"))
-    monkeypatch.setattr("autoskillit.planner.merge.write_versioned_json", spy)
-
-    src = tmp_path / "combined.json"
-    src.write_text(json.dumps({"phases": [{"id": "P1", "name": "Old"}], "schema_version": 1}))
-    repl = tmp_path / "replacement.json"
-    repl.write_text(json.dumps({"id": "P1", "name": "New"}))
-
-    result = replace_item(str(src), "P1", str(repl))
-    assert result["replaced_id"] == "P1"
-    data = json.loads(src.read_text())
-    assert data["phases"][0]["name"] == "New"
-    assert data["schema_version"] == 1
-    spy.assert_not_called()
-
-
-# --- expand functions propagate task context ---
-
-
-def test_expand_assignments_includes_task_file_path_in_context(tmp_path: Path) -> None:
-    from autoskillit.planner.manifests import expand_assignments
-
-    task_file = tmp_path / "task_input.md"
-    task_file.write_text("Split research.yaml into 4 sub-recipes")
-    refined = {
-        "task": "Split research.yaml into 4 sub-recipes",
-        "phases": [
-            {
-                "id": "P1",
-                "name": "Recipe Split",
-                "assignments_preview": [{"name": "Sub-recipe A"}, {"name": "Sub-recipe B"}],
-            }
-        ],
-    }
-    plan_path = tmp_path / "refined_plan.json"
-    plan_path.write_text(json.dumps(refined))
-    result = expand_assignments(
-        refined_plan_path=str(plan_path),
-        output_dir=str(tmp_path),
-        task_file_path=str(task_file),
-    )
-    context_paths = [p for p in result["context_paths"].split(",") if p.strip()]
-    assert context_paths, "Must produce at least one context path"
-    for cp in context_paths:
-        context = json.loads(Path(cp).read_text())
-        assert "task_file_path" in context, "Context file must include task_file_path field"
-        assert context["task_file_path"] == str(task_file)
-        assert "task" not in context, "Context file must not include inline task text"
-
-
-def test_expand_assignments_without_task_file_path_uses_empty(tmp_path: Path) -> None:
-    from autoskillit.planner.manifests import expand_assignments
-
-    refined = {
-        "task": "Some task",
-        "phases": [{"id": "P1", "name": "Phase", "assignments_preview": [{"name": "A"}]}],
-    }
-    plan_path = tmp_path / "refined_plan.json"
-    plan_path.write_text(json.dumps(refined))
-    result = expand_assignments(refined_plan_path=str(plan_path), output_dir=str(tmp_path))
-    context_paths = [p for p in result["context_paths"].split(",") if p.strip()]
-    assert context_paths, "Must produce at least one context path"
-    for cp in context_paths:
-        context = json.loads(Path(cp).read_text())
-        assert context.get("task_file_path", "") == ""
-        assert "task" not in context, "Context file must not include inline task text"
-
-
-def test_expand_wps_includes_task_file_path_in_context(tmp_path: Path) -> None:
-    from autoskillit.planner.manifests import expand_wps
-
-    task_file = tmp_path / "task_input.md"
-    task_file.write_text("Add telemetry to fleet dispatch")
-    refined = {
-        "schema_version": 2,
-        "task": "Add telemetry to fleet dispatch",
-        "source_dir": "/fake",
-        "assignments": [
-            {
-                "id": "P1-A1",
-                "phase_id": "P1",
-                "phase_number": 1,
-                "assignment_number": 1,
-                "name": "Telemetry",
-                "goal": "Add metrics",
-                "scope": ["fleet"],
-                "deliverables": ["fleet/_api.py"],
-                "proposed_work_packages": [
-                    {
-                        "id": "P1-A1-WP1",
-                        "name": "WP: Instrument dispatch",
-                        "scope": "fleet/_api.py",
-                        "estimated_files": ["fleet/_api.py"],
-                    }
-                ],
-            }
-        ],
-    }
-    plan_path = tmp_path / "refined_assignments.json"
-    plan_path.write_text(json.dumps(refined))
-    result = expand_wps(
-        refined_assignments_path=str(plan_path),
-        output_dir=str(tmp_path),
-        task_file_path=str(task_file),
-    )
-    context_paths = [p for p in result["context_paths"].split(",") if p.strip()]
-    assert context_paths, "Must produce at least one context path"
-    for cp in context_paths:
-        context = json.loads(Path(cp).read_text())
-        assert "task_file_path" in context, "Context file must include task_file_path field"
-        assert context["task_file_path"] == str(task_file)
-        assert "task" not in context, "Context file must not include inline task text"
-
-
-# --- ID Contract Enforcement Tests (1a, 1b, 1c, 1g) ---
-
-
-def test_expand_assignments_normalizes_plain_string_previews(tmp_path: Path) -> None:
-    """Test 1a: plain-string previews are normalized to canonical P{N}-A{M} IDs."""
-    from autoskillit.planner.manifests import expand_assignments
-
-    refined = tmp_path / "refined_plan.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "phases": [
-                    {
-                        "id": "P3",
-                        "name": "Phase 3",
-                        "ordering": 3,
-                        "assignments_preview": ["WP-1: Define Backend", "WP-2: Implement API"],
-                    },
-                ]
-            }
-        )
-    )
-    result = expand_assignments(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    assignment_ids = manifest["items"][0]["metadata"]["assignment_ids"]
-    assignment_names = manifest["items"][0]["metadata"]["assignment_names"]
-    assert assignment_ids == ["P3-A1", "P3-A2"], (
-        f"Expected canonical IDs ['P3-A1', 'P3-A2'], got {assignment_ids}"
-    )
-    assert assignment_names == ["WP-1: Define Backend", "WP-2: Implement API"], (
-        f"Raw string names should be preserved in assignment_names, got {assignment_names}"
-    )
-
-
-def test_expand_assignments_normalizes_dict_with_noncanonical_id(tmp_path: Path) -> None:
-    """Test 1b: dict previews with non-canonical id are normalized to canonical format."""
-    from autoskillit.planner.manifests import expand_assignments
-
-    refined = tmp_path / "refined_plan.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "phases": [
-                    {
-                        "id": "P1",
-                        "name": "Phase 1",
-                        "ordering": 1,
-                        "assignments_preview": [{"id": "setup-infra", "name": "Setup Infra"}],
-                    },
-                ]
-            }
-        )
-    )
-    result = expand_assignments(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    assignment_ids = manifest["items"][0]["metadata"]["assignment_ids"]
-    assert assignment_ids == ["P1-A1"], (
-        f"Expected canonical ['P1-A1'], got {assignment_ids} — non-canonical must be normalized"
-    )
-
-
-@pytest.mark.parametrize(
-    "preview_input,expected_ids",
-    [
-        # Plain strings with WP-style prefixes
-        (
-            ["WP-1: Task A", "WP-2: Task B"],
-            ["P1-A1", "P1-A2"],
-        ),
-        # Dicts with canonical IDs
-        (
-            [{"id": "P1-A1", "name": "A"}, {"id": "P1-A2", "name": "B"}],
-            ["P1-A1", "P1-A2"],
-        ),
-        # Dicts without IDs
-        (
-            [{"name": "First"}, {"name": "Second"}],
-            ["P1-A1", "P1-A2"],
-        ),
-        # Mixed: dict with canonical id + plain string
-        (
-            [{"id": "P1-A1", "name": "First"}, "Task B"],
-            ["P1-A1", "P1-A2"],
-        ),
-        # Dict with non-canonical id
-        (
-            [{"id": "setup-task", "name": "Setup"}],
-            ["P1-A1"],
-        ),
-    ],
-)
-def test_expand_assignments_post_generation_all_ids_match_regex(
-    tmp_path: Path, preview_input: list, expected_ids: list[str]
-) -> None:
-    """Test 1c: every generated assignment ID matches ASSIGN_ID_RE."""
-    from autoskillit.planner.manifests import expand_assignments
-    from autoskillit.planner.schema import ASSIGN_ID_RE
-
-    refined = tmp_path / "refined_plan.json"
-    refined.write_text(
-        json.dumps(
-            {
-                "phases": [
-                    {
-                        "id": "P1",
-                        "name": "Phase 1",
-                        "ordering": 1,
-                        "assignments_preview": preview_input,
-                    },
-                ]
-            }
-        )
-    )
-    result = expand_assignments(str(refined), str(tmp_path))
-    manifest = json.loads(Path(result["manifest_path"]).read_text())
-    assignment_ids = manifest["items"][0]["metadata"]["assignment_ids"]
-    assert assignment_ids == expected_ids, f"Expected {expected_ids}, got {assignment_ids}"
-    for aid in assignment_ids:
-        assert ASSIGN_ID_RE.match(aid), (
-            f"Assignment ID {aid!r} does not match canonical format ^P\\d+-A\\d+$"
-        )
-
-
-def test_validate_refined_plan_accepts_non_empty_plain_string_preview(tmp_path: Path) -> None:
-    """Test 1g: validate_refined_plan accepts plain-string entries that are non-empty strings."""
-    from autoskillit.planner.schema import validate_refined_plan
-
-    data = {
-        "phases": [
-            {
-                "id": "P1",
-                "name": "Phase 1",
-                "ordering": 1,
-                "assignments_preview": ["WP-1: Task A", {"id": "P1-A2", "name": "Task B"}],
-            },
-        ]
-    }
-    # Should not raise — plain strings are valid previews
-    result = validate_refined_plan(data)
-    assert result["phases"][0]["assignments_preview"] == data["phases"][0]["assignments_preview"]
