--- tests/planner/test_validation.py
@@ -1,1044 +1,0 @@
-"""Tests for validate_plan callable."""
-
-from __future__ import annotations
-
-import json
-from pathlib import Path
-
-import pytest
-
-from autoskillit.planner.schema import DELIVERABLE_BOUNDS
-from autoskillit.planner.validation import (
-    _check_assignment_completeness,
-    _check_dag_acyclic,
-    _check_duplicate_deliverables,
-    _check_phase_completeness,
-    _check_sizing_bounds,
-    _load_assignment_results,
-    _load_phase_results,
-    _load_wp_results,
-    validate_plan,
-)
-from tests.planner.conftest import (
-    make_assignment_result,
-    make_phase_result,
-    make_wp_result,
-    write_json,
-)
-
-pytestmark = [pytest.mark.layer("planner"), pytest.mark.small, pytest.mark.feature("planner")]
-
-
-# ---------------------------------------------------------------------------
-# Fixture helpers
-# ---------------------------------------------------------------------------
-
-
-def _make_minimal_output_dir(
-    tmp_path: Path,
-    *,
-    num_phases: int = 1,
-    wps_per_assignment: int = 1,
-    deliverables_override: list[str] | None = None,
-    depends_on_override: dict[str, list[str]] | None = None,
-    extra_phases: list[int] | None = None,
-    extra_assignments: list[tuple[int, int]] | None = None,
-) -> Path:
-    """Build a minimal valid output_dir structure."""
-    phases_dir = tmp_path / "phases"
-    assigns_dir = tmp_path / "assignments"
-    wps_dir = tmp_path / "work_packages"
-
-    for p in range(1, num_phases + 1):
-        write_json(
-            phases_dir / f"P{p}_result.json",
-            make_phase_result(p, name=f"Phase {p}"),
-        )
-
-    for p in range(1, num_phases + 1):
-        for a in range(1, 2):
-            write_json(
-                assigns_dir / f"P{p}-A{a}_result.json",
-                make_assignment_result(
-                    p,
-                    a,
-                    name=f"Assignment P{p}-A{a}",
-                    proposed_work_packages=[
-                        f"P{p}-A{a}-WP{w}" for w in range(1, wps_per_assignment + 1)
-                    ],
-                ),
-            )
-
-    for p in range(1, num_phases + 1):
-        for a in range(1, 2):
-            for w in range(1, wps_per_assignment + 1):
-                wp_id = f"P{p}-A{a}-WP{w}"
-                deliverables = (
-                    deliverables_override
-                    if deliverables_override is not None
-                    else [f"src/mod_{wp_id}.py"]
-                )
-                deps = (depends_on_override or {}).get(wp_id, [])
-                write_json(
-                    wps_dir / f"{wp_id}_result.json",
-                    make_wp_result(wp_id, deliverables=deliverables, depends_on=deps),
-                )
-
-    manifest_items = []
-    for p in range(1, num_phases + 1):
-        for a in range(1, 2):
-            for w in range(1, wps_per_assignment + 1):
-                manifest_items.append({"id": f"P{p}-A{a}-WP{w}", "status": "done"})
-    write_json(
-        wps_dir / "wp_manifest.json",
-        {"pass_name": "work_packages", "items": manifest_items},
-    )
-
-    if extra_phases:
-        for p in extra_phases:
-            write_json(
-                phases_dir / f"P{p}_result.json",
-                make_phase_result(p, name=f"Phase {p}"),
-            )
-
-    if extra_assignments:
-        for p, a in extra_assignments:
-            write_json(
-                assigns_dir / f"P{p}-A{a}_result.json",
-                make_assignment_result(p, a, name=f"Orphan assignment P{p}-A{a}"),
-            )
-
-    return tmp_path
-
-
-# ---------------------------------------------------------------------------
-# Tests
-# ---------------------------------------------------------------------------
-
-
-def test_validate_plan_valid_returns_pass(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    assert result["issue_count"] == "0"
-
-
-def test_validate_plan_cyclic_dep_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp1 = "P1-A1-WP1"
-    wp2 = "P1-A1-WP2"
-    wp_dir = tmp_path / "work_packages"
-    data1 = json.loads((wp_dir / f"{wp1}_result.json").read_text())
-    data1["depends_on"] = [wp2]
-    (wp_dir / f"{wp1}_result.json").write_text(json.dumps(data1))
-    data2 = json.loads((wp_dir / f"{wp2}_result.json").read_text())
-    data2["depends_on"] = [wp1]
-    (wp_dir / f"{wp2}_result.json").write_text(json.dumps(data2))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(wp1 in f["message"] and wp2 in f["message"] for f in validation["findings"])
-
-
-def test_validate_plan_missing_dep_ref_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, depends_on_override={"P1-A1-WP1": ["P1-A1-WP99"]})
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_phase_no_assignments_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, extra_phases=[2])
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_assignment_no_wps_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, extra_assignments=[(1, 2)])
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_wp_zero_deliverables_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    wp_path = tmp_path / "work_packages" / "P1-A1-WP1_result.json"
-    raw = json.loads(wp_path.read_text())
-    raw["deliverables"] = []
-    wp_path.write_text(json.dumps(raw))
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_wp_too_many_deliverables_warns_not_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    wp_path = tmp_path / "work_packages" / "P1-A1-WP1_result.json"
-    raw = json.loads(wp_path.read_text())
-    _, hi = DELIVERABLE_BOUNDS
-    raw["deliverables"] = [f"f{i}.py" for i in range(hi + 1)]
-    wp_path.write_text(json.dumps(raw))
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    assert int(result["warning_count"]) == 1
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert validation["warnings"][0]["check"] == "sizing_bounds"
-
-
-def test_validate_plan_duplicate_deliverables_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2, deliverables_override=["src/foo.py"])
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_failed_wp_flagged_but_not_sole_fail_cause(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    manifest_path = tmp_path / "work_packages" / "wp_manifest.json"
-    manifest = json.loads(manifest_path.read_text())
-    manifest["items"][0]["status"] = "failed"
-    manifest_path.write_text(json.dumps(manifest))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any("P1-A1-WP1" in f["message"] for f in validation["findings"])
-
-
-def test_validate_plan_dep_graph_backward_dep_injection(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    write_json(
-        tmp_path / "dep_graph.json",
-        {"added_backward_deps": {"P1-A1-WP2": ["P1-A1-WP1"]}, "forward_deps": {}},
-    )
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-
-
-def test_validate_plan_dep_graph_creates_cycle_fails(tmp_path: Path) -> None:
-    _make_minimal_output_dir(
-        tmp_path, wps_per_assignment=2, depends_on_override={"P1-A1-WP1": ["P1-A1-WP2"]}
-    )
-    write_json(
-        tmp_path / "dep_graph.json",
-        {"added_backward_deps": {"P1-A1-WP2": ["P1-A1-WP1"]}, "forward_deps": {}},
-    )
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-
-
-def test_validate_plan_writes_validation_json(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    result = validate_plan(str(tmp_path))
-    validation_path = tmp_path / "validation.json"
-    assert validation_path.exists()
-    data = json.loads(validation_path.read_text())
-    assert "verdict" in data
-    assert result["validation_path"] == str(validation_path)
-
-
-def test_validate_plan_return_values_are_strings(tmp_path: Path) -> None:
-    _make_minimal_output_dir(tmp_path)
-    result = validate_plan(str(tmp_path))
-    assert all(isinstance(v, str) for v in result.values())
-
-
-# ---------------------------------------------------------------------------
-# _check_duplicate_files_touched tests (T12–T14)
-# ---------------------------------------------------------------------------
-
-
-def test_check_duplicate_files_touched_detects_overlap() -> None:
-    """T12: Two WPs touching the same file produce a finding."""
-    from autoskillit.planner.validation import _check_duplicate_files_touched
-
-    wp_results = {
-        "P1-A1-WP1": {"files_touched": ["src/foo.py", "src/bar.py"]},
-        "P2-A1-WP1": {"files_touched": ["src/foo.py", "src/baz.py"]},
-    }
-    findings = _check_duplicate_files_touched(wp_results)
-    assert len(findings) == 1
-    assert "src/foo.py" in findings[0]["message"]
-    assert "P1-A1-WP1" in findings[0]["message"]
-    assert "P2-A1-WP1" in findings[0]["message"]
-
-
-def test_check_duplicate_files_touched_no_false_positives() -> None:
-    """T13: WPs with disjoint files_touched produce no findings."""
-    from autoskillit.planner.validation import _check_duplicate_files_touched
-
-    wp_results = {
-        "P1-A1-WP1": {"files_touched": ["src/foo.py"]},
-        "P1-A1-WP2": {"files_touched": ["src/bar.py"]},
-    }
-    findings = _check_duplicate_files_touched(wp_results)
-    assert findings == []
-
-
-def test_validate_plan_includes_duplicate_files_touched(tmp_path: Path) -> None:
-    """T14: validate_plan detects files_touched overlap as warning, not error."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    for wp_id in ("P1-A1-WP1", "P1-A1-WP2"):
-        result_path = wp_dir / f"{wp_id}_result.json"
-        data = json.loads(result_path.read_text())
-        data["files_touched"] = ["src/shared.py"]
-        result_path.write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any("src/shared.py" in w["message"] for w in validation["warnings"])
-
-
-# ---------------------------------------------------------------------------
-# Severity-level tests (T15–T19)
-# ---------------------------------------------------------------------------
-
-
-def test_warning_severity_does_not_fail_verdict(tmp_path: Path) -> None:
-    """T15: files_touched overlap is a warning, not an error — verdict stays pass."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    for wp_id in ("P1-A1-WP1", "P1-A1-WP2"):
-        result_path = wp_dir / f"{wp_id}_result.json"
-        data = json.loads(result_path.read_text())
-        data["files_touched"] = ["src/shared.py"]
-        result_path.write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    assert result["issue_count"] == "0"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert len(validation["findings"]) == 0
-    assert len(validation["warnings"]) == 1
-    assert validation["warnings"][0]["severity"] == "warning"
-    assert validation["warnings"][0]["check"] == "duplicate_files_touched"
-    assert "src/shared.py" in validation["warnings"][0]["message"]
-
-
-def test_error_findings_have_structured_fields(tmp_path: Path) -> None:
-    """T16: Error findings contain message, severity, and check fields."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp1 = "P1-A1-WP1"
-    wp2 = "P1-A1-WP2"
-    wp_dir = tmp_path / "work_packages"
-    data1 = json.loads((wp_dir / f"{wp1}_result.json").read_text())
-    data1["depends_on"] = [wp2]
-    (wp_dir / f"{wp1}_result.json").write_text(json.dumps(data1))
-    data2 = json.loads((wp_dir / f"{wp2}_result.json").read_text())
-    data2["depends_on"] = [wp1]
-    (wp_dir / f"{wp2}_result.json").write_text(json.dumps(data2))
-
-    validate_plan(str(tmp_path))
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    for finding in validation["findings"]:
-        assert "message" in finding
-        assert "severity" in finding
-        assert "check" in finding
-        assert finding["severity"] == "error"
-    cycle_findings = [f for f in validation["findings"] if f["check"] == "dag_acyclic"]
-    assert len(cycle_findings) == 1
-
-
-def test_mixed_errors_and_warnings(tmp_path: Path) -> None:
-    """T17: Sizing violation (error) + files_touched overlap (warning) coexist."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    for wp_id in ("P1-A1-WP1", "P1-A1-WP2"):
-        result_path = wp_dir / f"{wp_id}_result.json"
-        data = json.loads(result_path.read_text())
-        data["deliverables"] = []
-        data["files_touched"] = ["src/shared.py"]
-        result_path.write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    error_findings = [f for f in validation["findings"] if f["severity"] == "error"]
-    assert len(error_findings) >= 1
-    assert len(validation["warnings"]) == 1
-    assert validation["warnings"][0]["severity"] == "warning"
-
-
-def test_validation_json_schema_version_2(tmp_path: Path) -> None:
-    """T18: validation.json uses schema_version 2 and includes warnings key."""
-    _make_minimal_output_dir(tmp_path)
-    validate_plan(str(tmp_path))
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert validation["schema_version"] == 2
-    assert "warnings" in validation
-
-
-def test_check_duplicate_files_touched_returns_structured_findings() -> None:
-    """T19: _check_duplicate_files_touched returns dicts with message/severity/check."""
-    from autoskillit.planner.validation import _check_duplicate_files_touched
-
-    wp_results = {
-        "P1-A1-WP1": {"files_touched": ["src/foo.py"]},
-        "P2-A1-WP1": {"files_touched": ["src/foo.py"]},
-    }
-    findings = _check_duplicate_files_touched(wp_results)
-    assert len(findings) == 1
-    assert findings[0]["severity"] == "warning"
-    assert findings[0]["check"] == "duplicate_files_touched"
-    assert "message" in findings[0]
-
-
-# ---------------------------------------------------------------------------
-# _check_version_bump_steps tests (T-VB-1 through T-VB-8)
-# ---------------------------------------------------------------------------
-
-
-def test_version_bump_step_pyproject(tmp_path: Path) -> None:
-    """T-VB-1: WP with technical_steps referencing pyproject.toml version edit → warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["technical_steps"] = ["Edit pyproject.toml version field to X.Y.Z"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert validation["findings"] == []
-    assert any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_sync_versions_task(tmp_path: Path) -> None:
-    """T-VB-2: WP with technical_steps containing sync-versions → warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["technical_steps"] = ["Run task sync-versions"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_sync_versions_py(tmp_path: Path) -> None:
-    """T-VB-3: WP with technical_steps containing sync_versions.py → warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["technical_steps"] = ["python3 scripts/sync_versions.py"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_no_match(tmp_path: Path) -> None:
-    """T-VB-4: WP with unrelated technical_steps → no version_bump_step warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["technical_steps"] = ["Refactor the API handler"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert not any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_in_name(tmp_path: Path) -> None:
-    """T-VB-5: WP with version-bump pattern in name → warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["name"] = "WP3: Version Bump"
-    data["technical_steps"] = []
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_no_technical_steps_key(tmp_path: Path) -> None:
-    """T-VB-6: WP with no technical_steps at all → no crash, no warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data.pop("technical_steps", None)
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert not any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_version_bump_step_only_flagged_wp(tmp_path: Path) -> None:
-    """T-VB-7: Multiple WPs, only one triggers the warning → one entry per flagged WP."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    data1 = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data1["technical_steps"] = ["Edit pyproject.toml version field to 1.2.3"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data1))
-    data2 = json.loads((wp_dir / "P1-A1-WP2_result.json").read_text())
-    data2["technical_steps"] = ["Implement the feature"]
-    (wp_dir / "P1-A1-WP2_result.json").write_text(json.dumps(data2))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    vb_warnings = [w for w in validation["warnings"] if w["check"] == "version_bump_step"]
-    assert len(vb_warnings) == 1
-    assert "P1-A1-WP1" in vb_warnings[0]["message"]
-
-
-def test_version_bump_step_coexists_with_duplicate_files_touched(tmp_path: Path) -> None:
-    """T-VB-8: version_bump_step warning coexists with duplicate_files_touched warning."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    data1 = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data1["technical_steps"] = ["Edit pyproject.toml version field to 1.2.3"]
-    data1["files_touched"] = ["src/shared.py"]
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data1))
-    data2 = json.loads((wp_dir / "P1-A1-WP2_result.json").read_text())
-    data2["files_touched"] = ["src/shared.py"]
-    (wp_dir / "P1-A1-WP2_result.json").write_text(json.dumps(data2))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    checks = {w["check"] for w in validation["warnings"]}
-    assert "version_bump_step" in checks
-    assert "duplicate_files_touched" in checks
-    assert len(validation["findings"]) == 0
-
-
-def test_version_bump_step_via_summary(tmp_path: Path) -> None:
-    """T-VB-9: WP with version-bump pattern in summary only → warning."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    data = json.loads((wp_dir / "P1-A1-WP1_result.json").read_text())
-    data["summary"] = "Perform a version-bump before releasing."
-    data["technical_steps"] = []
-    (wp_dir / "P1-A1-WP1_result.json").write_text(json.dumps(data))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(w["check"] == "version_bump_step" for w in validation["warnings"])
-
-
-def test_validate_plan_warns_on_phase_sentinel_in_assignments_dir(tmp_path: Path) -> None:
-    """Phase-tier result files in assignments/ emit a file_discovery_miss warning."""
-    _make_minimal_output_dir(tmp_path)
-    sentinel = {"id": "P1", "status": "complete", "assignment_count": 1, "failed_count": 0}
-    write_json(tmp_path / "assignments" / "P1_result.json", sentinel)
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(
-        w["check"] == "file_discovery_miss" and w["severity"] == "warning"
-        for w in validation["warnings"]
-    )
-
-
-@pytest.mark.parametrize(
-    "bad_dep,reason",
-    [
-        ("P1-A1", "assignment-level ID in WP depends_on"),
-        ("P1", "phase-level ID in WP depends_on"),
-        ("WP1", "bare WP ID without phase/assignment prefix"),
-    ],
-)
-def test_check_dep_id_format_rejects_malformed(bad_dep, reason, tmp_path: Path) -> None:
-    """depends_on entries must match WP ID format (P\\d+-A\\d+-WP\\d+)."""
-    from autoskillit.planner.validation import _check_dep_id_format
-
-    wp_results = {
-        "P1-A1-WP1": {"id": "P1-A1-WP1", "depends_on": [bad_dep]},
-        "P1-A1-WP2": {"id": "P1-A1-WP2", "depends_on": []},
-    }
-    findings = _check_dep_id_format(wp_results)
-    assert any(f["severity"] == "error" and bad_dep in f["message"] for f in findings), (
-        f"Expected error for malformed dep {bad_dep!r}: {reason}"
-    )
-
-
-def test_validate_plan_warns_on_non_wp_result_file_in_work_packages_dir(tmp_path: Path) -> None:
-    """Non-WP result files in work_packages/ emit a file_discovery_miss warning."""
-    _make_minimal_output_dir(tmp_path)
-    sentinel = {"id": "P1", "status": "complete", "assignment_count": 1, "failed_count": 0}
-    write_json(tmp_path / "work_packages" / "P1_result.json", sentinel)
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(
-        w["check"] == "file_discovery_miss" and w["severity"] == "warning"
-        for w in validation["warnings"]
-    )
-
-
-def test_validate_plan_emits_discovery_miss_warning_for_non_matching_files(tmp_path: Path) -> None:
-    """When *_result.json files exist but don't match tier regex, a warning finding is emitted."""
-    _make_minimal_output_dir(tmp_path, num_phases=1, wps_per_assignment=1)
-    stray = {"id": "stray", "status": "complete", "assignment_count": 1, "failed_count": 0}
-    write_json(tmp_path / "work_packages" / "stray_result.json", stray)
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(
-        w["check"] == "file_discovery_miss" and w["severity"] == "warning"
-        for w in validation["warnings"]
-    ), "Expected a file_discovery_miss warning for stray_result.json"
-
-
-def test_discover_tier_files_returns_accepted_and_rejected(tmp_path: Path) -> None:
-    """discover_tier_files returns both accepted files and rejected (non-matching) files."""
-    from autoskillit.planner.schema import PHASE_RESULT_FILE_RE
-    from autoskillit.planner.validation import DiscoveryResult, discover_tier_files
-
-    phases_dir = tmp_path / "phases"
-    phases_dir.mkdir(parents=True)
-    write_json(phases_dir / "P1_result.json", make_phase_result(1))
-    stray = {"id": "stray", "status": "complete"}
-    write_json(phases_dir / "stray_result.json", stray)
-
-    result = discover_tier_files(phases_dir, PHASE_RESULT_FILE_RE)
-    assert isinstance(result, DiscoveryResult)
-    assert len(result.accepted) == 1
-    assert result.accepted[0].name == "P1_result.json"
-    assert len(result.rejected) == 1
-    assert result.rejected[0].name == "stray_result.json"
-
-
-def test_validate_plan_exempts_absorbed_assignments(tmp_path: Path) -> None:
-    """When all WPs of an assignment are absorbed, no false completeness error fires."""
-    phases_dir = tmp_path / "phases"
-    assigns_dir = tmp_path / "assignments"
-    wps_dir = tmp_path / "work_packages"
-
-    write_json(phases_dir / "P1_result.json", make_phase_result(1))
-    write_json(
-        assigns_dir / "P1-A1_result.json",
-        make_assignment_result(1, 1, proposed_work_packages=["P1-A1-WP1"]),
-    )
-    write_json(
-        assigns_dir / "P1-A2_result.json",
-        make_assignment_result(1, 2, proposed_work_packages=["P1-A2-WP1"]),
-    )
-    write_json(
-        wps_dir / "P1-A1-WP1_result.json",
-        make_wp_result("P1-A1-WP1", deliverables=["src/a.py"]),
-    )
-    write_json(
-        wps_dir / "wp_manifest.json",
-        {"pass_name": "work_packages", "items": [{"id": "P1-A1-WP1", "status": "done"}]},
-    )
-    registry = {
-        "schema_version": 1,
-        "absorbed": {"P1-A2-WP1": {"merged_into": "P1-A1-WP1", "group_id": "P1-A1-WP1"}},
-    }
-    write_json(wps_dir / "absorption_registry.json", registry)
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert not any(
-        f["check"] == "assignment_completeness" and "P1-A2" in f["message"]
-        for f in validation["findings"]
-    ), "Assignment P1-A2 should be exempt due to absorption registry"
-    assert not any(
-        f["check"] == "assignment_completeness" and "P1-A2" in f["message"]
-        for f in validation.get("warnings", [])
-    ), "Assignment P1-A2 should not appear as a warning either"
-
-
-# ---------------------------------------------------------------------------
-# Direct unit tests for _check_sizing_bounds
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.parametrize(
-    ("count", "expected_findings", "expected_severity"),
-    [
-        (0, 1, "error"),
-        (DELIVERABLE_BOUNDS[0], 0, None),
-        (5, 0, None),
-        (DELIVERABLE_BOUNDS[1] + 1, 1, "warning"),
-    ],
-)
-def test_check_sizing_bounds_boundary_values(
-    count: int, expected_findings: int, expected_severity: str | None
-) -> None:
-    wp_results = {
-        "P1-A1-WP1": {"deliverables": [f"f{i}.py" for i in range(count)]},
-    }
-    findings = _check_sizing_bounds(wp_results)
-    assert len(findings) == expected_findings
-    if expected_findings:
-        assert findings[0]["check"] == "sizing_bounds"
-        assert findings[0]["severity"] == expected_severity
-
-
-def test_check_sizing_bounds_uses_deliverable_bounds_constant() -> None:
-    lo, hi = DELIVERABLE_BOUNDS
-    wp_at_lo = {"P1-A1-WP1": {"deliverables": [f"f{i}.py" for i in range(lo)]}}
-    wp_at_hi = {"P1-A1-WP2": {"deliverables": [f"f{i}.py" for i in range(hi)]}}
-    assert _check_sizing_bounds(wp_at_lo) == []
-    assert _check_sizing_bounds(wp_at_hi) == []
-
-
-def test_check_sizing_bounds_upper_violation_is_warning() -> None:
-    _, hi = DELIVERABLE_BOUNDS
-    wp_results = {
-        "P1-A1-WP1": {"deliverables": [f"f{i}.py" for i in range(hi + 1)]},
-    }
-    findings = _check_sizing_bounds(wp_results)
-    assert len(findings) == 1
-    assert findings[0]["severity"] == "warning"
-    assert findings[0]["check"] == "sizing_bounds"
-
-
-# ---------------------------------------------------------------------------
-# Direct unit tests for _check_duplicate_deliverables
-# ---------------------------------------------------------------------------
-
-
-def test_check_duplicate_deliverables_detects_shared() -> None:
-    wp_results = {
-        "P1-A1-WP1": {"deliverables": ["src/shared.py", "src/a.py"]},
-        "P1-A1-WP2": {"deliverables": ["src/shared.py", "src/b.py"]},
-    }
-    findings = _check_duplicate_deliverables(wp_results)
-    assert len(findings) == 1
-    assert findings[0]["check"] == "duplicate_deliverables"
-    assert "src/shared.py" in findings[0]["message"]
-
-
-def test_check_duplicate_deliverables_no_false_positives() -> None:
-    wp_results = {
-        "P1-A1-WP1": {"deliverables": ["src/a.py"]},
-        "P1-A1-WP2": {"deliverables": ["src/b.py"]},
-    }
-    findings = _check_duplicate_deliverables(wp_results)
-    assert findings == []
-
-
-def test_check_duplicate_deliverables_returns_structured_findings() -> None:
-    wp_results = {
-        "P1-A1-WP1": {"deliverables": ["src/shared.py"]},
-        "P1-A1-WP2": {"deliverables": ["src/shared.py"]},
-    }
-    findings = _check_duplicate_deliverables(wp_results)
-    assert len(findings) == 1
-    f = findings[0]
-    assert "message" in f
-    assert f["severity"] == "error"
-    assert f["check"] == "duplicate_deliverables"
-
-
-# ---------------------------------------------------------------------------
-# Deduplication-to-orphan cascade
-# ---------------------------------------------------------------------------
-
-
-def test_deduplication_orphan_cascade(tmp_path: Path) -> None:
-    """Simulates dedup removing WP-B's only deliverable → sizing_bounds error."""
-    _make_minimal_output_dir(tmp_path, wps_per_assignment=2)
-    wp_dir = tmp_path / "work_packages"
-    wp_b = json.loads((wp_dir / "P1-A1-WP2_result.json").read_text())
-    wp_b["deliverables"] = []
-    (wp_dir / "P1-A1-WP2_result.json").write_text(json.dumps(wp_b))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "fail"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    sizing_findings = [f for f in validation["findings"] if f["check"] == "sizing_bounds"]
-    assert len(sizing_findings) == 1
-    assert "P1-A1-WP2" in sizing_findings[0]["message"]
-
-
-def test_validate_plan_reads_finalized_manifest(tmp_path: Path) -> None:
-    """validate_plan must find wp_manifest.json written by finalize_wp_manifest."""
-    from autoskillit.planner import finalize_wp_manifest
-
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    finalize_wp_manifest(str(wp_dir), str(tmp_path))
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-
-
-# ---------------------------------------------------------------------------
-# Silent-drop elimination — non-canonical filenames must raise, not silently drop
-# ---------------------------------------------------------------------------
-
-
-def test_discover_tier_files_places_non_canonical_in_rejected(tmp_path: Path) -> None:
-    """Non-canonical *_result.json files end up in the rejected list, not accepted."""
-    from autoskillit.planner.schema import WP_RESULT_FILE_RE
-    from autoskillit.planner.validation import discover_tier_files
-
-    wp_dir = tmp_path / "work_packages"
-    wp_dir.mkdir()
-    write_json(wp_dir / "P1-A1-WP1_result.json", make_wp_result("P1-A1-WP1"))
-    write_json(wp_dir / "P1-A1-WP2a_result.json", {"id": "P1-A1-WP2a", "name": "bad"})
-
-    result = discover_tier_files(wp_dir, WP_RESULT_FILE_RE)
-    assert [f.name for f in result.accepted] == ["P1-A1-WP1_result.json"]
-    assert [f.name for f in result.rejected] == ["P1-A1-WP2a_result.json"]
-
-
-def test_load_wp_results_returns_non_canonical_in_rejected(tmp_path: Path) -> None:
-    """_load_wp_results returns non-canonical filenames in the rejected list."""
-    wp_dir = tmp_path / "work_packages"
-    wp_dir.mkdir()
-    write_json(wp_dir / "P1-A1-WP1_result.json", make_wp_result("P1-A1-WP1"))
-    write_json(wp_dir / "P1-A1-WP2a_result.json", {"id": "P1-A1-WP2a", "name": "bad"})
-
-    results, rejected = _load_wp_results(tmp_path)
-    assert "P1-A1-WP1" in results
-    assert len(rejected) == 1
-    assert rejected[0].name == "P1-A1-WP2a_result.json"
-
-
-def test_load_phase_results_returns_non_canonical_in_rejected(tmp_path: Path) -> None:
-    """_load_phase_results returns non-canonical filenames in the rejected list."""
-    phases_dir = tmp_path / "phases"
-    phases_dir.mkdir()
-    write_json(phases_dir / "P1_result.json", make_phase_result(1))
-    write_json(phases_dir / "Phase1_result.json", {"id": "P1", "name": "bad"})
-
-    results, rejected = _load_phase_results(tmp_path)
-    assert "P1" in results
-    assert len(rejected) == 1
-    assert rejected[0].name == "Phase1_result.json"
-
-
-def test_load_assignment_results_returns_non_canonical_in_rejected(tmp_path: Path) -> None:
-    """_load_assignment_results returns non-canonical filenames in the rejected list."""
-    assigns_dir = tmp_path / "assignments"
-    assigns_dir.mkdir()
-    write_json(
-        assigns_dir / "P1-A1_result.json",
-        make_assignment_result(1, 1, name="Good"),
-    )
-    write_json(
-        assigns_dir / "P1-A2b_result.json", {"id": "P1-A2b", "phase_id": "P1", "name": "bad"}
-    )
-
-    results, rejected = _load_assignment_results(tmp_path)
-    assert "P1-A1" in results
-    assert len(rejected) == 1
-    assert rejected[0].name == "P1-A2b_result.json"
-
-
-def test_validate_plan_emits_warning_for_non_canonical_wp_file(tmp_path: Path) -> None:
-    """validate_plan emits a file_discovery_miss warning for a non-canonical WP file."""
-    _make_minimal_output_dir(tmp_path)
-    wp_dir = tmp_path / "work_packages"
-    write_json(wp_dir / "P1-A1-WP2a_result.json", {"id": "P1-A1-WP2a", "name": "bad"})
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert any(
-        w["check"] == "file_discovery_miss" and "P1-A1-WP2a_result.json" in w["message"]
-        for w in validation["warnings"]
-    ), "Expected file_discovery_miss warning for P1-A1-WP2a_result.json"
-
-
-# ---------------------------------------------------------------------------
-# DAG cycle decomposition tests
-# ---------------------------------------------------------------------------
-
-
-def test_check_dag_acyclic_decomposes_2_node_cycle() -> None:
-    """2-node mutual cycles are reported with cycle_size=2 and edges."""
-    wp_results = {
-        "P1-A1-WP1": {"depends_on": ["P1-A1-WP2"]},
-        "P1-A1-WP2": {"depends_on": ["P1-A1-WP1"]},
-        "P1-A1-WP3": {"depends_on": []},
-    }
-    findings = _check_dag_acyclic(wp_results)
-    assert len(findings) == 1
-    assert findings[0]["cycle_size"] == 2
-    assert set(findings[0]["cycle_nodes"]) == {"P1-A1-WP1", "P1-A1-WP2"}
-    assert set(map(tuple, findings[0]["cycle_edges"])) == {
-        ("P1-A1-WP1", "P1-A1-WP2"),
-        ("P1-A1-WP2", "P1-A1-WP1"),
-    }
-
-
-def test_check_dag_acyclic_3_node_cycle_no_edges() -> None:
-    """3+ node cycles report cycle_size and cycle_nodes but no edges (not safe to auto-fix)."""
-    wp_results = {
-        "P1-A1-WP1": {"depends_on": ["P1-A1-WP3"]},
-        "P1-A1-WP2": {"depends_on": ["P1-A1-WP1"]},
-        "P1-A1-WP3": {"depends_on": ["P1-A1-WP2"]},
-    }
-    findings = _check_dag_acyclic(wp_results)
-    assert len(findings) == 1
-    assert findings[0]["cycle_size"] == 3
-    assert set(findings[0]["cycle_nodes"]) == {"P1-A1-WP1", "P1-A1-WP2", "P1-A1-WP3"}
-    assert "cycle_edges" not in findings[0]
-
-
-# ---------------------------------------------------------------------------
-# Lifecycle registry tests (Tests 1B, 1C, 1E, 1G)
-# ---------------------------------------------------------------------------
-
-
-def test_check_assignment_completeness_skips_voided_assignments() -> None:
-    """1B: _check_assignment_completeness skips voided assignments from lifecycle registry."""
-    assignment_results = {
-        "P4-A2": {"phase_number": 4, "assignment_number": 2},
-        "P4-A1": {"phase_number": 4, "assignment_number": 1},
-        "P4-A3": {"phase_number": 4, "assignment_number": 3},
-    }
-    wp_results = {
-        "P4-A1-WP1": {"id": "P4-A1-WP1", "depends_on": []},
-    }
-    lifecycle_registry = {
-        "voided_assignments": ["P4-A2"],
-        "voided_phases": [],
-        "absorbed": {},
-    }
-    findings = _check_assignment_completeness(assignment_results, wp_results, lifecycle_registry)
-    assert not any("P4-A2" in f["message"] for f in findings)
-    assert any("P4-A3" in f["message"] for f in findings)
-
-
-def test_check_phase_completeness_skips_voided_phases() -> None:
-    """1C: _check_phase_completeness skips voided phases from lifecycle registry."""
-    phase_results = {
-        "P3": {"phase_number": 3},
-        "P1": {"phase_number": 1},
-        "P2": {"phase_number": 2},
-    }
-    assignment_results = {
-        "P1-A1": {"phase_number": 1, "assignment_number": 1},
-    }
-    lifecycle_registry = {
-        "voided_phases": ["P3"],
-        "voided_assignments": [],
-        "absorbed": {},
-    }
-    findings = _check_phase_completeness(phase_results, assignment_results, lifecycle_registry)
-    assert not any("P3" in f["message"] for f in findings)
-    assert any("P2" in f["message"] for f in findings)
-
-
-def test_validate_plan_voided_assignment_no_false_positive(tmp_path: Path) -> None:
-    """1E: End-to-end — voided assignment does NOT produce validation error."""
-    from autoskillit.planner.merge import merge_refined_assignments
-
-    phases_dir = tmp_path / "phases"
-    assigns_dir = tmp_path / "assignments"
-    wps_dir = tmp_path / "work_packages"
-
-    write_json(phases_dir / "P1_result.json", make_phase_result(1))
-    write_json(
-        assigns_dir / "P1-A1_result.json",
-        make_assignment_result(1, 1, proposed_work_packages=["P1-A1-WP1"]),
-    )
-    write_json(
-        assigns_dir / "P1-A2_result.json",
-        make_assignment_result(1, 2, proposed_work_packages=["P1-A2-WP1"]),
-    )
-    write_json(
-        wps_dir / "P1-A1-WP1_result.json",
-        make_wp_result("P1-A1-WP1", deliverables=["src/a.py"]),
-    )
-    write_json(
-        wps_dir / "wp_manifest.json",
-        {"pass_name": "work_packages", "items": [{"id": "P1-A1-WP1", "status": "done"}]},
-    )
-
-    ctx_dir = tmp_path / "refine_contexts"
-    ctx_dir.mkdir()
-    voided_assignment = {
-        "id": "P1-A2",
-        "phase_id": "P1",
-        "name": "Voided",
-        "goal": "g",
-        "technical_approach": "",
-        "proposed_work_packages": [],
-    }
-    normal_assignment = {
-        "id": "P1-A1",
-        "phase_id": "P1",
-        "name": "Normal",
-        "goal": "g",
-        "technical_approach": "",
-        "proposed_work_packages": [
-            {
-                "id": "P1-A1-WP1",
-                "name": "WP1",
-                "summary": "s",
-                "goal": "g",
-                "technical_steps": [],
-                "files_touched": ["src/a.py"],
-                "apis_defined": [],
-                "apis_consumed": [],
-                "depends_on": [],
-                "deliverables": ["src/a.py"],
-                "acceptance_criteria": [],
-            }
-        ],
-    }
-    write_json(
-        ctx_dir / "P1_result.json",
-        {"schema_version": 1, "assignments": [normal_assignment, voided_assignment]},
-    )
-
-    merge_refined_assignments(planner_dir=str(tmp_path))
-    result = validate_plan(str(tmp_path))
-
-    assert result["verdict"] == "pass"
-    validation = json.loads((tmp_path / "validation.json").read_text())
-    assert not any(
-        f["check"] == "assignment_completeness" and "P1-A2" in f["message"]
-        for f in validation["findings"]
-    )
-
-
-def test_validate_plan_backward_compat_absorption_registry(tmp_path: Path) -> None:
-    """1G: validate_plan reads absorption_registry.json when lifecycle_registry.json is absent."""
-    phases_dir = tmp_path / "phases"
-    assigns_dir = tmp_path / "assignments"
-    wps_dir = tmp_path / "work_packages"
-
-    write_json(phases_dir / "P1_result.json", make_phase_result(1))
-    write_json(
-        assigns_dir / "P1-A1_result.json",
-        make_assignment_result(1, 1, proposed_work_packages=["P1-A1-WP1"]),
-    )
-    write_json(
-        assigns_dir / "P1-A2_result.json",
-        make_assignment_result(1, 2, proposed_work_packages=["P1-A2-WP1"]),
-    )
-    write_json(
-        wps_dir / "P1-A1-WP1_result.json",
-        make_wp_result("P1-A1-WP1", deliverables=["src/a.py"]),
-    )
-    write_json(
-        wps_dir / "wp_manifest.json",
-        {"pass_name": "work_packages", "items": [{"id": "P1-A1-WP1", "status": "done"}]},
-    )
-    registry = {
-        "schema_version": 1,
-        "absorbed": {"P1-A2-WP1": {"merged_into": "P1-A1-WP1", "group_id": "P1-A1-WP1"}},
-    }
-    write_json(wps_dir / "absorption_registry.json", registry)
-
-    result = validate_plan(str(tmp_path))
-    assert result["verdict"] == "pass"
-    assert not any(
-        "P1-A2" in f.get("message", "")
-        for f in json.loads((tmp_path / "validation.json").read_text())["findings"]
-    )
